Send up sfJobOpp with minimal lists.
Show case in short candidate source display
